Fixed issue with Texture loading on Windows. Dwitched STB Image load call.

+++ src/debug_io.h
@@ -6,7 +6,7 @@
 
 struct debug_read_file_result
 {
-    u32 ContentSize;
+    u64 ContentSize;
     void *Content;
 };
